refactor: move pooling layers to pooling.py
All pooling layers are moved to file `layers/pooling.py`.
